refactor: /research/centers SSR로 수정
